#149: validation integration
