+++ src/integrations/validation/errors/UnknownImplementation.ts
@@ -1,0 +1,10 @@
+
+import ValidationError from './ValidationError.js';
+
+export default class UnknownImplementation extends ValidationError
+{
+    constructor(name?: string) 
+    {
+        super(`Unknown validation implementation: ${name}`);
+    }
+}
