+++ src/integrations/validation/errors/UnknownValidator.ts
@@ -1,0 +1,10 @@
+
+import ValidationError from './ValidationError.js';
+
+export default class UnknownValidator extends ValidationError
+{
+    constructor(name?: string) 
+    {
+        super(`Unknown validator: ${name}`);
+    }
+}
