+++ src/integrations/validation/errors/ValidationError.ts
@@ -1,0 +1,8 @@
+
+export default class ValidationError extends Error
+{
+    constructor(message?: string)
+    {
+        super(message ?? 'Validation error');
+    }
+}
