+++ src/integrations/validation/module.ts
@@ -1,6 +1,8 @@
 
 export * from './definitions/constants.js';
-export * from './definitions/errors.js';
 export * from './definitions/types.js';
+export * from './errors/InvalidData.js';
+export * from './errors/UnknownImplementation.js';
+export * from './errors/UnknownValidator.js';
+export * from './errors/ValidationError.js';
 export { default } from './implementation.js';
-
